requirements.txt
fastapi
uvicorn
jinja2
python-frontmatter
markdown
pandas
python-dotenv
google-generativeaifirebase-admin
